Werror stuff, minor n64 changes

+++ vendored/str_hash.hxx
@@ -9,7 +9,7 @@
 constexpr uint32_t str_hash(std::string_view data) noexcept {
     uint32_t hash = 5381;
     const size_t size = data.size();
-    for(int i = 0; i < size; i++)
+    for(size_t i = 0; i < size; i++)
         hash = ((hash << 5) + hash) + (unsigned char)data[i];
 
     return hash;
